+++ server/.gitignore
@@ -1,3 +1,4 @@
 __pycache__/
 *.egg-info/
 .venv/
+backend/__pycache__
